Initial structure of the array
The initial structure of the array is
defined, which will contain the map.
The idea is that each sector is made up
of a defining object, material and
color. And another object that defines
what material is in that sector.
It was implemented to show the labels,
representing each tile, according to
item and tile

Resolve: T#002