Cursor, status bar, delete items.
Added custom cursor to make it more
pleasant, and the crosshair changes
to a hand when aiming at a pick-up item.
When we pick up the item, a window
appears showing the picked up item.
Also, the energy bars were made and the
mechanic that items can be consumed to
recover these bars was created. And the
possibility of deleting items was added
generating a UI that asks if you are
sure to delete.

Resolve: T#001