Quick Slots (UI, Scripts basic)
Code refactoring for cursor visibility,
depending on the situation and to make
it work in the executable version of
the game. Implemented that when opening
the inventory or crafting menu, you
cannot interact with terrain objects
and the crosshair is not visible.
Creation of Frame_with_border.png and
creation of the Quick_Slots UI. When
right-clicking with the mouse, if an
item is equipable, it moves to the next
empty equipment slot. You can also drag
the item from the inventory to equipment
slots and vice versa. Pressing keys 1
to 7 selects an equipment slot. If it
has something, it is highlighted and
that item cannot be dragged. If you
choose the slot that is already chosen,
it is deselected.

Resolve: T#002_partial