Quick Slot, Model tool and animation
Selecting the corresponding slot calls
a new method, which instantiates the
tool so that it is visible to the
player. An Axe model was downloaded and
the resource was generated, defining
position and rotation. Animations for
when it is in idle state and action were
created. The animation was disabled in
certain situations, such as when the
inventory is open. A camera was added
that shows the axe (equippable layer),
so that it does not disappear behind trees

Resolve: T#002